+++ packages/backend/src/services/CatalogService/CatalogService.ts
@@ -10,16 +10,20 @@ import {
     CatalogTable,
     CatalogType,
     ChartSummary,
+    DEFAULT_METRICS_EXPLORER_TIME_INTERVAL,
     Explore,
     ExploreError,
+    FieldType,
     ForbiddenError,
+    getAvailableTimeDimensionsFromTables,
+    getDefaultTimeDimension,
     hasIntersection,
     InlineErrorType,
     isExploreError,
     MAX_METRICS_TREE_NODE_COUNT,
+    MetricWithAssociatedTimeDimension,
     NotFoundError,
     ParameterError,
-    parseMetricsTreeNodeId,
     SessionUser,
     SummaryExplore,
     TablesConfiguration,
@@ -39,7 +43,6 @@ import { LightdashAnalytics } from '../../analytics/LightdashAnalytics';
 import { LightdashConfig } from '../../config/parseConfig';
 import type {
     DbCatalogTagsMigrateIn,
-    DbMetricsTreeEdge,
     DbMetricsTreeEdgeIn,
 } from '../../database/entities/catalog';
 import { CatalogModel } from '../../models/CatalogModel/CatalogModel';
@@ -403,33 +406,40 @@ export class CatalogService<
         const currentCatalogItems =
             await this.catalogModel.getCatalogItemsSummary(projectUuid);
 
-        const metricEdgesMigrateIn: DbMetricsTreeEdgeIn[] = prevMetricTreeEdges
-            .filter((edge): edge is CatalogMetricsTreeEdge => {
+        const metricEdgesMigrateIn: DbMetricsTreeEdgeIn[] =
+            prevMetricTreeEdges.reduce<DbMetricsTreeEdgeIn[]>((acc, edge) => {
                 const sourceCatalogItem = currentCatalogItems.find(
                     (catalogItem) =>
                         catalogItem.name === edge.source.name &&
-                        catalogItem.tableName === edge.source.tableName,
+                        catalogItem.tableName === edge.source.tableName &&
+                        catalogItem.type === CatalogType.Field &&
+                        catalogItem.fieldType === FieldType.METRIC,
                 );
 
                 const targetCatalogItem = currentCatalogItems.find(
                     (catalogItem) =>
                         catalogItem.name === edge.target.name &&
-                        catalogItem.tableName === edge.target.tableName,
+                        catalogItem.tableName === edge.target.tableName &&
+                        catalogItem.type === CatalogType.Field &&
+                        catalogItem.fieldType === FieldType.METRIC,
                 );
 
-                return Boolean(sourceCatalogItem) && Boolean(targetCatalogItem);
-            })
-            .map((edge) => ({
-                source_metric_name: edge.source.name,
-                source_metric_table_name: edge.source.tableName,
-                source_metric_type: CatalogType.Field,
-                target_metric_name: edge.target.name,
-                target_metric_table_name: edge.target.tableName,
-                target_metric_type: CatalogType.Field,
-                project_uuid: edge.projectUuid,
-                created_by_user_uuid: edge.createdByUserUuid,
-                created_at: edge.createdAt,
-            }));
+                if (sourceCatalogItem && targetCatalogItem) {
+                    return [
+                        ...acc,
+                        {
+                            source_metric_catalog_search_uuid:
+                                sourceCatalogItem.catalogSearchUuid,
+                            target_metric_catalog_search_uuid:
+                                targetCatalogItem.catalogSearchUuid,
+                            created_by_user_uuid: edge.createdByUserUuid,
+                            created_at: edge.createdAt,
+                        },
+                    ];
+                }
+
+                return acc;
+            }, []);
 
         return this.catalogModel.migrateMetricsTreeEdges(metricEdgesMigrateIn);
     }
@@ -873,7 +883,7 @@ export class CatalogService<
         projectUuid: string,
         tableName: string,
         metricName: string,
-    ) {
+    ): Promise<MetricWithAssociatedTimeDimension> {
         const { organizationUuid } = await this.projectModel.getSummary(
             projectUuid,
         );
@@ -902,83 +912,118 @@ export class CatalogService<
         }
 
         const filteredExplore = getFilteredExplore(explore, userAttributes);
-
-        const metric =
-            filteredExplore?.tables?.[tableName]?.metrics?.[metricName];
-
-        // Get the default time dimension from the metric, or the explore if not set on the metric
-        const defaultTimeDimension =
-            metric.defaultTimeDimension ??
-            filteredExplore?.tables?.[tableName]?.defaultTimeDimension;
+        const tables = filteredExplore?.tables;
+        const metric = tables?.[tableName]?.metrics?.[metricName];
+        const metricBaseTable = tables?.[metric?.table];
 
         if (!metric) {
             throw new NotFoundError('Metric not found');
         }
 
+        const defaultTimeDimension = getDefaultTimeDimension(
+            metric,
+            metricBaseTable,
+        );
+
+        let availableTimeDimensions:
+            | ReturnType<typeof getAvailableTimeDimensionsFromTables>
+            | undefined;
+
+        // If no default time dimension is defined, we can use the available time dimensions so the user can see what time dimensions are available
+        if (!defaultTimeDimension) {
+            availableTimeDimensions =
+                getAvailableTimeDimensionsFromTables(tables);
+        }
+
+        let timeDimension:
+            | MetricWithAssociatedTimeDimension['timeDimension']
+            | undefined;
+
+        if (defaultTimeDimension) {
+            timeDimension = {
+                field: defaultTimeDimension.field,
+                interval: defaultTimeDimension.interval,
+                table: metric.table,
+            };
+        } else if (
+            availableTimeDimensions &&
+            availableTimeDimensions.length > 0
+        ) {
+            const firstAvailableTimeDimension = availableTimeDimensions[0];
+
+            if (!firstAvailableTimeDimension.isIntervalBase) {
+                throw new Error(
+                    'The first available time dimension is not an interval base dimension',
+                );
+            }
+
+            timeDimension = {
+                field: firstAvailableTimeDimension.name,
+                interval: DEFAULT_METRICS_EXPLORER_TIME_INTERVAL,
+                table: firstAvailableTimeDimension.table,
+            };
+        }
+
         return {
             ...metric,
-            defaultTimeDimension,
+            ...(availableTimeDimensions && availableTimeDimensions.length > 0
+                ? { availableTimeDimensions }
+                : {}),
+            timeDimension,
         };
     }
 
     async getMetricsTree(
         user: SessionUser,
         projectUuid: string,
-        // Using metricIds instead of filters and searching metrics because we might want to have the ability to select specific metrics rather than filter by catalog tags
-        metricIds: string[],
+        metricUuids: string[],
     ) {
         // TODO: check permissions
-        if (metricIds.length > MAX_METRICS_TREE_NODE_COUNT) {
+        if (metricUuids.length > MAX_METRICS_TREE_NODE_COUNT) {
             throw new ParameterError(
                 `Cannot get more than ${MAX_METRICS_TREE_NODE_COUNT} metrics in the metrics tree`,
             );
         }
 
-        return this.catalogModel.getMetricsTree(
-            projectUuid,
-            metricIds.map((id) => parseMetricsTreeNodeId(id)),
-        );
+        return this.catalogModel.getMetricsTree(projectUuid, metricUuids);
     }
 
     async createMetricsTreeEdge(
         user: SessionUser,
         projectUuid: string,
-        { sourceMetricId, targetMetricId }: ApiMetricsTreeEdgePayload,
+        {
+            sourceCatalogSearchUuid,
+            targetCatalogSearchUuid,
+        }: ApiMetricsTreeEdgePayload,
     ) {
         // TODO: check permissions
-        const edgeSource = parseMetricsTreeNodeId(sourceMetricId);
-        const edgeTarget = parseMetricsTreeNodeId(targetMetricId);
-
-        const sourceCatalogItem = await this.catalogModel.getCatalogItemByName(
-            projectUuid,
-            edgeSource.name,
-            edgeSource.tableName,
-            CatalogType.Field,
+        const sourceCatalogItem = await this.catalogModel.getCatalogItem(
+            sourceCatalogSearchUuid,
         );
+
+        if (sourceCatalogItem?.field_type !== FieldType.METRIC) {
+            throw new ParameterError('Source metric is not a valid metric');
+        }
 
         if (!sourceCatalogItem) {
             throw new NotFoundError('Source metric not found');
         }
 
-        const targetCatalogItem = await this.catalogModel.getCatalogItemByName(
-            projectUuid,
-            edgeTarget.name,
-            edgeTarget.tableName,
-            CatalogType.Field,
+        const targetCatalogItem = await this.catalogModel.getCatalogItem(
+            targetCatalogSearchUuid,
         );
 
         if (!targetCatalogItem) {
             throw new NotFoundError('Target metric not found');
         }
 
+        if (targetCatalogItem?.field_type !== FieldType.METRIC) {
+            throw new ParameterError('Target metric is not a valid metric');
+        }
+
         return this.catalogModel.createMetricsTreeEdge({
-            source_metric_name: edgeSource.name,
-            source_metric_table_name: edgeSource.tableName,
-            source_metric_type: CatalogType.Field,
-            target_metric_name: edgeTarget.name,
-            target_metric_table_name: edgeTarget.tableName,
-            target_metric_type: CatalogType.Field,
-            project_uuid: projectUuid,
+            source_metric_catalog_search_uuid: sourceCatalogSearchUuid,
+            target_metric_catalog_search_uuid: targetCatalogSearchUuid,
             created_by_user_uuid: user.userUuid,
         });
     }
@@ -986,18 +1031,15 @@ export class CatalogService<
     deleteMetricsTreeEdge(
         user: SessionUser,
         projectUuid: string,
-        { sourceMetricId, targetMetricId }: ApiMetricsTreeEdgePayload,
+        {
+            sourceCatalogSearchUuid,
+            targetCatalogSearchUuid,
+        }: ApiMetricsTreeEdgePayload,
     ) {
         // TODO: check permissions
-        const edgeSource = parseMetricsTreeNodeId(sourceMetricId);
-        const edgeTarget = parseMetricsTreeNodeId(targetMetricId);
-
         return this.catalogModel.deleteMetricsTreeEdge({
-            source_metric_name: edgeSource.name,
-            source_metric_table_name: edgeSource.tableName,
-            target_metric_name: edgeTarget.name,
-            target_metric_table_name: edgeTarget.tableName,
-            project_uuid: projectUuid,
+            source_metric_catalog_search_uuid: sourceCatalogSearchUuid,
+            target_metric_catalog_search_uuid: targetCatalogSearchUuid,
         });
     }
 }