+++ packages/common/src/types/metricsExplorer.ts
@@ -1,3 +1,4 @@
+import { type MetricWithAssociatedTimeDimension } from './catalog';
 import type { ItemsMap } from './field';
 import type { ResultRow } from './results';
 
@@ -7,14 +8,22 @@ export enum MetricExplorerComparison {
     DIFFERENT_METRIC = 'different_metric',
 }
 
-export type MetricExplorerDateRange = [Date | null, Date | null];
+export type MetricExplorerPartialDateRange = [Date | null, Date | null];
+export type MetricExplorerDateRange = [Date, Date];
 
 export type MetricExplorerComparisonType =
     | { type: MetricExplorerComparison.NONE }
     | { type: MetricExplorerComparison.PREVIOUS_PERIOD }
     | { type: MetricExplorerComparison.DIFFERENT_METRIC; metricName: string };
 
+export type MetricExploreDataPoint = {
+    date: Date;
+    metric: unknown;
+    compareMetric: unknown;
+};
+
 export type MetricsExplorerQueryResults = {
+    metric: MetricWithAssociatedTimeDimension;
     rows: ResultRow[];
     comparisonRows: ResultRow[] | undefined;
     fields: ItemsMap;